pyproject.toml
[project]
name = "haya"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Bhargav Kulkarni", email = "bhargav.kulkarni1@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "docutils>=0.21.2",
    "libsass>=0.23.0",
    "mako>=1.3.8",
    "pydantic>=2.10.5",
]

[project.scripts]
haya = "haya:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pylsp]
pythonPath = "./.venv"  # Adjust for your `uv` environment
